Prepare browser preview launch validation
Add launch docs, landing assets, GitHub Pages and Playwright smoke tooling. Gate catalog thumbnails through the asset manifest, add web smoke hooks, and fix CanvasKit shader compatibility. Tighten export/Android behavior and document the reference corpus boundary.

+++ shaders/cellular_and_stochastic/wireworld_gpu.frag
@@ -96,7 +96,7 @@ void main() {
   vec2 p = uv / max(0.000001, uZoom) + uCenter;
 
   int target = int(clamp(uIterations, 1.0, float(MAX_ITERS)));
-  int steps = clamp(target / 6, 1, 80);
+  int steps = int(clamp(float(target / 6), 1.0, 80.0));
   ivec2 cell = ivec2(floor(p * 96.0));
 
   int state = evolveCell(cell, steps);
